Change:hart, irq, cpuid, plic_{} types to usize

+++ kernel-rs/src/memlayout.rs
@@ -23,7 +23,7 @@ pub const UART0_IRQ: usize = 10;
 
 /// virtio mmio interface
 pub const VIRTIO0: usize = 0x10001000;
-pub const VIRTIO0_IRQ: i32 = 1;
+pub const VIRTIO0_IRQ: usize = 1;
 
 /// local interrupt controller, which contains the timer.
 pub const CLINT: usize = 0x2000000;
@@ -39,17 +39,17 @@ pub const CLINT_MTIME: usize = CLINT.wrapping_add(0xbff8);
 /// qemu puts programmable interrupt controller here.
 pub const PLIC: usize = 0xc000000;
 pub const PLIC_PENDING: usize = PLIC.wrapping_add(0x1000);
-pub const fn plic_senable(hart: i32) -> usize {
+pub const fn plic_senable(hart: usize) -> usize {
     PLIC.wrapping_add(0x2080)
-        .wrapping_add((hart as usize).wrapping_mul(0x100))
+        .wrapping_add((hart).wrapping_mul(0x100))
 }
-pub const fn plic_spriority(hart: i32) -> usize {
+pub const fn plic_spriority(hart: usize) -> usize {
     PLIC.wrapping_add(0x201000)
-        .wrapping_add((hart as usize).wrapping_mul(0x2000))
+        .wrapping_add((hart).wrapping_mul(0x2000))
 }
-pub const fn plic_sclaim(hart: i32) -> usize {
+pub const fn plic_sclaim(hart: usize) -> usize {
     PLIC.wrapping_add(0x201004)
-        .wrapping_add((hart as usize).wrapping_mul(0x2000))
+        .wrapping_add((hart).wrapping_mul(0x2000))
 }
 
 /// the kernel expects there to be RAM
